# api/openapi.yaml
openapi: 3.0.3
info:
  title: Maigo URL Shortener API
  version: 0.1.0
  description: |
    Maigo is a terminal-first URL shortener backed by PostgreSQL.

    The CLI uses the OAuth 2.0 authorization-code flow with mandatory PKCE
    (S256). Access tokens are JWTs. Refresh tokens are rotated and persisted
    server-side so logout and revocation take effect.
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v1/urls:
    post:
      tags: [URLs]
      summary: Create a short URL
      description: Requires an access token. The target accepts http and https URLs.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateURLRequest' }
      responses:
        '201':
          description: Short URL created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/URLMetadata' }
        '400':
          description: Invalid URL or expiration
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: Short-code conflict
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/v1/user/urls:
    get:
      tags: [URLs]
      summary: List the current user’s URLs
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: page_size
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: Paginated URL list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/URLListResponse' }
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/v1/urls/{code}:
    parameters:
      - name: code
        in: path
        required: true
        schema: { type: string, pattern: '^[0-9A-Za-z]{1,50}$' }
    get:
      tags: [URLs]
      summary: Get public URL metadata
      responses:
        '200':
          description: URL metadata
          content:
            application/json:
              schema: { $ref: '#/components/schemas/URLMetadata' }
        '400':
          description: Invalid short code
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: Short URL not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      tags: [URLs]
      summary: Delete an owned URL
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: URL deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: URL deleted successfully }
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: URL belongs to another user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: Short URL not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string, example: bad_request }
        message: { type: string, example: Invalid request parameters }
        details:
          nullable: true
          description: Optional validation details
...
    CreateURLRequest:
      type: object
      required: [url]
      properties:
        url: { type: string, format: uri, example: https://example.com/long/path }
        custom: { type: string, minLength: 3, maxLength: 50, pattern: '^[A-Za-z0-9]+$' }
        ttl: { type: integer, format: int64, minimum: 60, description: Lifetime in seconds }
        expires_at: { type: string, format: date-time, description: Alternative to ttl }
    URL:
      type: object
      properties:
        id: { type: integer, format: int64 }
        short_code: { type: string }
        target_url: { type: string, format: uri }
        created_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time, nullable: true }
        hits: { type: integer, format: int64 }
        user_id: { type: integer, format: int64, nullable: true }
    URLMetadata:
      allOf:
        - $ref: '#/components/schemas/URL'
        - type: object
          properties:
            url: { type: string, format: uri, description: Backward-compatible alias for target_url }
            short_url: { type: string, format: uri }
            expired: { type: boolean }
            expires_in: { type: integer, format: int64 }
    URLListResponse:
      type: object
      properties:
        urls:
          type: array
          items: { $ref: '#/components/schemas/URL' }
        pagination:
          type: object
          properties:
            page: { type: integer }
            page_size: { type: integer }
            total: { type: integer, format: int64 }
            pages: { type: integer }
...